Refactor semantic validators into their own package, and provide a base class that finds them and runs them.
